requirements.txt
blinker==1.4
click==7.1.2
Flask==1.1.2
Flask-Mail==0.9.1
Flask-MySQLdb==0.2.0
fpdf==1.7.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql==0.0.2
mysql-connector==2.2.9
mysqlclient==2.0.3
python-dotenv==0.19.2
Werkzeug==1.0.1
xlwt==1.3.0